iterate li in Paper.js
